add thread pool which only logs not handles message
